refactor : input, output 클래스명 수정
