+++ src/main/java/calculator/view/handler/InputHandler.java
@@ -1,0 +1,6 @@
+package calculator.view.handler;
+
+public interface InputHandler {
+
+    String inputValue();
+}
