--- src/main/java/calculator/view/input/InputView.java
@@ -1,6 +1,0 @@
-package calculator.view.input;
-
-public interface InputView {
-
-    String inputValue();
-}
